restructuring app less than 12 hours before presentation
