+++ package.json
@@ -12,28 +12,28 @@
   },
   "private": true,
   "dependencies": {
-    "ng2-img-cropper": "^0.9.0",
-    "@types/jquery": "^3.2.15",
-    "@cloudinary/angular-4.x": "^1.0.0",
-    "cloudinary-core": "^2.3.0",
-    "ngx-cookie-service": "^1.0.9",
     "@angular/animations": "^4.0.0",
+    "@angular/cdk": "^2.0.0-beta.12",
     "@angular/common": "^4.0.0",
     "@angular/compiler": "^4.0.0",
     "@angular/core": "^4.0.0",
     "@angular/forms": "^4.0.0",
     "@angular/http": "^4.0.0",
+    "@angular/material": "^2.0.0-beta.12",
     "@angular/platform-browser": "^4.0.0",
     "@angular/platform-browser-dynamic": "^4.0.0",
     "@angular/router": "^4.0.0",
-    "@angular/material": "^2.0.0-beta.12",
-    "@angular/cdk": "^2.0.0-beta.12",
+    "@cloudinary/angular-4.x": "^1.0.0",
+    "@types/jquery": "^3.2.15",
+    "cloudinary-core": "^2.3.0",
     "core-js": "^2.4.1",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4",
     "hammerjs": "^2.0.8",
+    "ng2-img-cropper": "^0.9.0",
+    "ngx-avatar": "^2.8.0",
+    "ngx-cookie-service": "^1.0.9",
+    "rxjs": "^5.1.0",
     "time-ago-pipe": "^1.2.1",
-    "ngx-avatar": "^2.8.0"
+    "zone.js": "^0.8.4"
   },
   "devDependencies": {
     "@angular/cli": "^1.4.8",
