+++ src/app/app.component.html
@@ -1,52 +1,1 @@
-<div class="everything">
-  <mat-toolbar color="primary">
-    <button mat-icon-button (click)="toggleSidenav()">
-      <mat-icon class="mat-24" aria-label="Menu">menu</mat-icon>
-    </button>
-    <span class="space-left">The Dank Network</span>
-    <span class="fill-space"></span>
-
-
-    <button mat-icon-button [matMenuTriggerFor]="notifications" (click)="readNotifications()">
-      <mat-icon *ngIf="this.notificationCount == 0" class="mat-24" matTooltip="Notifications"
-                aria-label="Notifications">notifications
-      </mat-icon>
-      <div *ngIf="this.notificationCount > 0" class="notificationCircle">{{this.notificationCount}}</div>
-      <mat-menu style="width: 80vw; max-width: 400px;" #notifications="matMenu">
-        <div class="notificationHolder">
-
-          <div class="noNotifications" *ngIf="notificationsList.length === 0">
-            <div class="noNotifText">No notifications here, senpai</div>
-          </div>
-          <mat-card *ngFor="let notification of notificationsList; let i = index" (click)="openUri(notification.uri)" class="notificationCard">
-            <mat-card-title class="notificationCardTitle">{{notification.fromUser.name}}</mat-card-title>
-            <div class="notificationCardTitle" style="float: right;">
-              <button style="line-height: 28px; height: auto; width: 26px;" (click)="deleteNotif(i, $event)" mat-icon-button color="warn">
-                <mat-icon aria-label="Delete notification">clear</mat-icon>
-              </button>
-            </div>
-            <mat-card-subtitle>{{notification.message}}</mat-card-subtitle>
-            <mat-card-subtitle>{{notification.date | timeAgo}}</mat-card-subtitle>
-          </mat-card>
-        </div>
-      </mat-menu>
-    </button>
-    <div style="margin-left: 0.5vh; cursor: pointer;">
-      <ngx-avatar #avatar routerLink="/profile" [size]="40" [textSizeRatio]="2" [name]="this.authenticatedUser.name" [src]="this.authenticatedUser.avatar" [initialsSize]="1"></ngx-avatar>
-    </div>
-  </mat-toolbar>
-  <mat-sidenav-container class="content-container">
-    <mat-sidenav #sidenav class="sidenav-menu">
-      <mat-nav-list>
-        <a *ngFor="let url of urls" mat-list-item (click)="disposeSidenav()" [routerLink]="[url.path]">
-          <mat-icon>{{url.icon}}</mat-icon>
-          <span style="width: 20px;"></span>{{url.desc}}
-          <span style="width: 20px;"></span>
-        </a>
-      </mat-nav-list>
-    </mat-sidenav>
-    <div class="routerOutlet">
-      <router-outlet></router-outlet>
-    </div>
-  </mat-sidenav-container>
-</div>
+<router-outlet></router-outlet>
