+++ src/app/apps/show-app-dialog/show-app-dialog.component.ts
@@ -37,4 +37,10 @@ export class ShowAppDialogComponent implements OnInit {
     });
   }
 
+  onKeyPress($event) {
+    if ($event.keyCode === 13) {
+      this.request();
+    }
+  }
+
 }
