+++ src/app/login/login.component.html
@@ -1,7 +1,7 @@
 <br>
 <div>
   <mat-card class="login-form">
-    <h3>Sign In</h3>
+    <h3>The Dank Network - Sign In</h3>
     <form>
       <mat-form-field class="example-full-width">
         <input matInput placeholder="User Name" [(ngModel)]="this.userLogin.username" name="inputUserName">
