+++ src/app/people/people.component.html
@@ -1,19 +1,19 @@
 <div class="wrapper">
 <nav mat-tab-nav-bar>
   <a mat-tab-link
-     [routerLink]="['find']"
+     [routerLink]="['/find']"
      routerLinkActive #rla="routerLinkActive"
      [active]="rla.isActive">
     FIND PEOPLE
   </a>
   <a mat-tab-link
-     [routerLink]="['followers']"
+     [routerLink]="['/followers']"
      routerLinkActive #rla2="routerLinkActive"
      [active]="rla2.isActive">
     FOLLOWERS
   </a>
   <a mat-tab-link
-     [routerLink]="['following']"
+     [routerLink]="['/following']"
      routerLinkActive #rla3="routerLinkActive"
      [active]="rla3.isActive">
     FOLLOWING
