+++ src/app/post/post.component.html
@@ -18,8 +18,8 @@
   </mat-card>
 </div>
 
-<div class="posts">
-  <mat-card *ngFor="let post of posts" class="postcard">
+<div class="posts grid">
+  <mat-card *ngFor="let post of posts" class="postcard griditem">
     <mat-card-header>
       <ngx-avatar mat-card-avatar size="40" src="{{post.user.avatar}}" initialsSize="1" routerLink="/profile/{{post.user.id}}" name={{post.user.name}}></ngx-avatar>
       <mat-card-title>{{post.user.name}}</mat-card-title>
