+++ src/app/register/register.component.ts
@@ -5,6 +5,7 @@ import {Http} from '@angular/http';
 import {MatSnackBar} from '@angular/material';
 import {Constants} from '../shared/constants';
 import {UserService} from '../user/user.service';
+import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-register',
@@ -16,16 +17,17 @@ export class RegisterComponent implements OnInit {
 
   userRegister: User = new User();
 
-  constructor (private userService: UserService, private snackBar: MatSnackBar) {
+  constructor (private userService: UserService, private snackBar: MatSnackBar, private router: Router) {
   }
 
   register() {
     this.userService.registerUser(this.userRegister).subscribe(user => {
       this.snackBar.open('Registered user successfully', 'Dismiss', {duration: 2000});
+      this.router.navigate(['/login']);
       console.log('Registred user:', user);
     }, (error: any) => {
       this.snackBar.open('User already exists', 'Dismiss', {duration: 2000});
-      console.log('Already register');
+      console.log('Already registered');
     });
   }
