+++ src/app/shared/constants.ts
@@ -2,4 +2,3 @@ export class Constants {
   public static url = 'http://localhost:8080/api';
   public static notificationUpdateTime = 15000;
 }
-
