+++ src/index.html
@@ -11,6 +11,7 @@
 <body>
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
   <script src="https://widget.cloudinary.com/global/all.js" type="text/javascript"></script>
+  <script src="https://unpkg.com/masonry-layout@4/dist/masonry.pkgd.min.js"></script>
   <script type="text/javascript">
   </script>
   <app-root></app-root>
