+++ src/styles.css
@@ -3,6 +3,7 @@
 body {
   margin: 0;
   height: 100%;
+  background: #fafafa;
 }
 html {
   height: 100%;
